requirements.txt
appdirs==1.4.3
asgiref==3.3.1
attrs==20.3.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
html5lib==1.0.1
idna==2.8
inflection==0.5.1
iniconfig==1.1.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.3
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
mypy-extensions==0.4.3
packaging==20.3
pathspec==0.8.1
pep517==0.8.2
pluggy==0.13.1
progress==1.5
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.6
python-decouple==3.4
pytoml==0.1.21
pytz==2021.1
regex==2021.3.17
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
